feat(timesheets): implement filtering and feature module
